Release initial Darn and Darnit work
This is just all the work done internally at Kusari by myself over
the past two weeks or so.

This includes work on Darn and Darnit.

Darn is a tool intended for running security remediation actions.
Actions are defined through yaml configuration.

Darnit is a tool intended to string together multiple darn actions
into workflows to remediate one or more findings as discovered in
an arbitrary report. These workflows end up in mapping files that
map findings to subworkflows or actions that should be run.